requirements.txt
torch==2.2.0+cu118
torchvision==0.17.0+cu118
torchaudio==2.2.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

numpy<2.0
dominate
visdom
imageio
matplotlib
pydantic
PyYAML
opencv-python